feat: complete u64 migration + smart_http error enum
u64 migration (all remaining fields):
- FileChunkRecord.range_start/range_end: u32→u64 (index/record.rs)
- ReconstructionChunkRange.start/end: u32→u64 (xet_adapter/model.rs)
- XorbObjectInfoV1: num_chunks, offsets, lengths all u32→u64
- XorbObject.info_length: u32→u64
- XorbObject format V1→V2 with backward compat in deserialize
- All callers updated across 19 files

smart_http error enum:
- Created SmartHttpError with 34 thiserror variants
- Replaced 11 map_err(format!(...)) + 26 other string errors
- Updated 5 function signatures from Result<_, String> to Result<_, SmartHttpError>
- 520 hub_api tests pass

7,963 tests pass, CI green

+++ crates/bench/benches/reconstruction.rs
@@ -25,8 +25,8 @@ fn make_record(chunk_count: usize) -> FileRecord {
             hash,
             offset,
             length,
-            range_start: i as u32,
-            range_end: i as u32 + 1,
+            range_start: i as u64,
+            range_end: i as u64 + 1,
             packed_start: offset,
             packed_end: offset + length,
         });
